fix php warnings

+++ centreon/www/include/configuration/configObject/host/DB-Func.php
@@ -344,12 +344,13 @@ function removeRelationLastHostDependency(int $hostId): void
               WHERE dependency_dep_id = (SELECT dependency_dep_id FROM dependency_hostParent_relation
                                          WHERE host_host_id =  ' . $hostId . ') GROUP BY dependency_dep_id';
     $dbResult = $pearDB->query($query);
-    $result = $dbResult->fetch();
 
-    //is last parent
-    if ($result['nb_dependency'] == 1) {
-        $pearDB->query("DELETE FROM dependency WHERE dep_id = " . $result['id']);
-    }
+    if (false !== ($result = $dbResult->fetch())) {
+        //is last parent
+        if ($result['nb_dependency'] == 1) {
+            $pearDB->query("DELETE FROM dependency WHERE dep_id = " . $result['id']);
+        }
+    };
 }
 
 function deleteHostInDB($hosts = array())
